# backend/src/main/java/com/acme/investment/interfaces/api/AuthController.java
package com.acme.investment.interfaces.api;

import com.acme.investment.infrastructure.persistence.UserEntity;
import com.acme.investment.infrastructure.persistence.UserJpaRepository;
import com.acme.investment.infrastructure.security.JwtService;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/auth")
public class AuthController {
    private final UserJpaRepository users;
    private final PasswordEncoder passwordEncoder;
    private final AuthenticationManager authenticationManager;
    private final JwtService jwtService;

    public AuthController(UserJpaRepository users, PasswordEncoder passwordEncoder,
                          AuthenticationManager authenticationManager, JwtService jwtService) {
        this.users = users;
        this.passwordEncoder = passwordEncoder;
        this.authenticationManager = authenticationManager;
        this.jwtService = jwtService;
    }

    @PostMapping("/register")
    ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        if (users.existsByEmail(request.email())) {
            return ResponseEntity.badRequest().build();
        }
        UserEntity user = users.save(new UserEntity(
                request.email().toLowerCase(),
                passwordEncoder.encode(request.password()),
                request.fullName()));
        return ResponseEntity.ok(new AuthResponse(jwtService.createToken(user.getEmail()), "Bearer"));
    }

    @PostMapping("/login")
    ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
        authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(request.email(), request.password()));
        return ResponseEntity.ok(new AuthResponse(jwtService.createToken(request.email()), "Bearer"));
    }

    record RegisterRequest(@Email String email, @Size(min = 12, max = 128) String password,
                           @NotBlank @Size(min = 2, max = 160) String fullName) {
    }

    record LoginRequest(@Email String email, @NotBlank String password) {
    }

    record AuthResponse(String token, String tokenType) {
    }
}