# backend/src/main/resources/openapi/investment-api.yaml
openapi: 3.0.3
info:
  title: Investment Portfolio Management API
  version: 0.1.0
  description: REST API for portfolio, watchlist, goal, news, and AI analysis workflows.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      security: []
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: JWT authentication response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/login:
    post:
      security: []
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT authentication response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/portfolios:
    get:
      summary: List current user's portfolios
      responses:
        '200':
          description: Portfolio list
    post:
      summary: Create a portfolio
      responses:
        '201':
          description: Portfolio created
  /api/v1/news:
    get:
      summary: List sourced news summaries
      responses:
        '200':
          description: News summaries include source, publication date, summary, and link.
  /api/v1/portfolios/{portfolioId}/analysis:
    post:
      summary: Generate sourced AI portfolio risk analysis
      parameters:
        - in: path
          name: portfolioId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis with citations
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        fullName:
          type: string
          minLength: 2
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, tokenType]
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer